refactor(cli): move routeMatcher into its own module
